src/requirements.txt
xlrd
sty
google-api-python-client
google-auth
google-auth-oauthlib